Support for percentages in property expression.
The different length properties all implement a Length interface
and all clients of the property subsystem must use the Length interface
when retrieving lengths.

The two different numerics (absolute and relative) both implement a
Numeric interface and all property calculations are done in terms of
the Numeric interface.

Lengths can be cast to numeric and can thus participate in expresions and
numerics can be cast to length (when dimension is 1) and can thus be used
as a length by the clients.

The call to Length.getValue() will cause relative lengths to be resolved
against their base length, and care must be taken to ensure that
getLength() is only called after all baselength has been assigned to the
FO elements.

PR: 26778 (first part).


git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/fop/trunk@197380 13f79535-47bb-0310-9956-ffa450edef68