1. Output constant types (RENDER_PDF, RENDER_PS, etc.) made common between CommandLineOptions and Driver, and factored into fo.Constants interface.
2. New Driver(AWTRenderer renderer) constructor added in apps.Driver for
the AWTRenderer (which does reloading, unique among the output types.)  Note
reloading currently doesn't work--AWTRenderer not yet functional.

3. Driver.hasData() method removed from API--a search on when it was implemented
dates it to 2001, when reset() was coded.  Unsure of its need, and so
removed until we have user requests for it.

4. Renderers abstracted away from API in favor of just calling .setRenderer(int Rendertype),
and configuring user-configurable options in FOUserAgent.

5. Driver.setRenderer(String <renderer class name>) also dropped.  If external user
need for it, can be re-implemented fairly easily via a "renderer override" string in FOUserAgent, for which FOTreeHandler can read and use instead.

6. Validity checking added to fo:flow, and convenience functions defining the "%block;"
parameter entity and neutral containers as defined in 6.2 of spec [1] added to FObj.

[1] http://www.w3.org/TR/2001/REC-xsl-20011015/slice6.html#section-N9447-Formatting-Object-Content


git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/fop/trunk@197745 13f79535-47bb-0310-9956-ffa450edef68