1.) Moved the SAXParser-creating method from FOFileHandler to CommandLineOptions, the latter being the class directly using it.
2.) Decoupled the SVG classes from apps.FOFileHandler, the SVG classes for some reason
want a string containing a name of a SAXParser.  We'll need to look in the future for
the actual need for this--normally error messages are done by the caller, not
the callee.  It would appear that Batik can generate a SAXParser name by default,
and use that when not explicitly overridden by the caller.


git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/fop/trunk@197804 13f79535-47bb-0310-9956-ffa450edef68