PR: Obtained from: Submitted by: Reviewed by: Start of work to move FOText from extending FObj to extending FONode instead.
All work is done except for changing FObj's parent from FObj to
FONode.  The RTF library works fine with the switch, but simple.fo and
franklin_alt.fo (among others I guess) will not work properly when FOText's
parent is switched.  (Layout problems occur--lines don't break properly
with the switch.)

Switch to super() constructor (instead of super(fobj) one) within
within TextLayoutManager, as it is not needed for this subclass, also to
facilitate the change above.


git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/fop/trunk@198040 13f79535-47bb-0310-9956-ffa450edef68