Implemented cloning and reparenting of the subtree under a marker to the retrieving retrieve-marker. The FO nodes are cloned. The new subtree is attached to the RetrieveMarker object. The property lists are not cloned. The subtree of property lists is attached to the property list of the RetrieveMarker object. This is only needed for the binding of the FO nodes. After that a subsequent retrieve-marker may reparent the property lists.
Removed Marker's reference to its property list. In Marker.endOfNode()
the parent of the property lists of the direct children of marker is
reset to null. FOText nodes are not mapped to the Marker's property
list in Marker. This allows the Marker's property list and its
ancestors to be GC'ed.

FOText nodes use the parent FO node's property list or the
RetrieveMarker's property list for binding.

TODO: Block.handleWhiteSpace is not applied to text in retrieve-marker.
TODO: White space treatment around retrieve-marker is incorrect.


git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/fop/trunk@198204 13f79535-47bb-0310-9956-ffa450edef68