Bugzilla #36505: Several fixes for SVG painting in the Java2DRenderer. Updated SVG test case to test additional features. Submitted by: Richard Wheeldon <richardw.at.geoquip-rnd.demon.co.uk>
Comments and changes by jeremias:
Added an additional SVG image with an intrinsic size and modified the test case to use and test that, too.
I'll commit this patch because it improves the current situation but I'm not happy with the current state of Java2DRenderer.renderSVGDocument(). Creating inverse transformation matrices is surely not the right way to do this. The matrix should be saved and restored instead.
Furthermore, renderSVGDocument should be extracted into a Java2DSVGHandler like it is done for the other renderers.

git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/fop/trunk@278966 13f79535-47bb-0310-9956-ffa450edef68