Fixed a memory-leak: The FO tree part of a page-sequence was not released when a page-sequence was finished.
Added a "MemoryEater" debug helper which can replicate page-sequences of a regular FO file using the fo-replicator.xsl. Without the fix a JVM set to the normal 64MB heap size throws an OutOfMemoryError when replicating readme.fo 150 times. With the fix the memory consumption doesn't go above 10MB for the 1350 pages. Note that about one third of the time goes into garbage collection.


git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/fop/trunk@428450 13f79535-47bb-0310-9956-ffa450edef68