[FEATURE] Adds Configurator to AbstractSVGDocumentHandler.
Beforehand, the SVGDocumentHandler could not be used directly and had to
be overwritten in the way, how the getConfigurator() method is over-
written. Also fonts could not be embedded without the configurator.

The foUserAgent can now be modified in the way:
foUserAgent.setRendererOverride(new SVGRenderer(foUserAgent));
foUserAgent.setDocumentHandlerOverride(new SVGDocumentHandler(IFContext context);

After that, the FopFactory.newFop(..) can be called and the transformation
can be done.